tests: move `move_to_next_cycle` to `DashTestFramework`
