+++ test/functional/feature_llmq_is_migration.py
@@ -6,7 +6,7 @@ import time
 
 from test_framework.messages import CTransaction, FromHex, hash256, ser_compact_size, ser_string
 from test_framework.test_framework import DashTestFramework
-from test_framework.util import wait_until, connect_nodes, sync_blocks
+from test_framework.util import wait_until, connect_nodes
 
 '''
 feature_llmq_is_migration.py
@@ -117,21 +117,6 @@ class LLMQISMigrationTest(DashTestFramework):
         for n in self.nodes:
             assert not n.quorum("hasrecsig", 104, request_id2, txid2)
 
-
-    def move_to_next_cycle(self):
-        cycle_length = 24
-        mninfos_online = self.mninfo.copy()
-        nodes = [self.nodes[0]] + [mn.node for mn in mninfos_online]
-        cur_block = self.nodes[0].getblockcount()
-
-        # move forward to next DKG
-        skip_count = cycle_length - (cur_block % cycle_length)
-        if skip_count != 0:
-            self.bump_mocktime(1, nodes=nodes)
-            self.nodes[0].generate(skip_count)
-        sync_blocks(nodes)
-        time.sleep(1)
-        self.log.info('Moved from block %d to %d' % (cur_block, self.nodes[0].getblockcount()))
 
 if __name__ == '__main__':
     LLMQISMigrationTest().main()
